git aliases: Add do, pwd, state, and stsh
In this commit I add several really useful aliases to Git. First is do,
which will run its command in the toplevel of the git project (where .git
is located). This will be handy for a fish function I want, bdo, which will
run commands from within the build-debug directory.

Second is pwd, which prints the relative directory from the toplevel to
CWD. This will be an important part of the "state" message.

Next is state, which will print a message containing the current branch,
the relative PWD, the message from the last commit, and the hash of the
last commit. The idea is to provide enough breadcrumbs to properly locate
that "state" within both time and space around a git project. Time, in the
sense of _where_ the "state" exists, relative to the commit history. Space,
in the sense of the correct branch and directory in which that "state"
was captured.

Finally is the stsh command, which is a specialized stash push, where the
push message is the output of `git state` in the CWD. This means that, when
I run git stsh, my changes are marked with a "state" and stashed. That way,
when I want to get back to work, I know what branch to check out, what
directory to cd to, and which commits I must deal with in order to get my
stashed-away work back into flow.