git: Improve stash shortcuts
In this commit, I improve the stash shortcuts, making state produce JSON,
and making the sl "stash list" command produce JSONL output, so I can query
it. Finally, I add the sr "stash relevant" which lists thouse stashes that
have the same branch and CWD as where the command is run. This should make
it pretty easy to keep track of those stashes which are most likely to be
relevant to the user at the time the command is run. Of course, the full
list is always available via either of "stash list" or "sr" (for JSONL)