Fix VPN configuration
In this commit, I finally fix the VPN configuration so that it
automatically adds the DNS crap, search domain crap, and reads auth info
from a file. Now I don't have to play around with keyrings or scripts. I
can just run openvpn as a daemon and it just works.